feat(photo-upload): added jury and user entry view page, photo full-screen on click
