+++ javinukai-front/src/pages/UserManagementPage.jsx
@@ -1,7 +1,7 @@
 import { useQuery } from "@tanstack/react-query";
 import { useState } from "react";
 import getUsers from "../services/users/getUsers";
-import PaginationSettings from "../Components/user-management/PaginationSettings";
+import PaginationSettings from "../Components/PaginationSettings";
 import { UserListItem } from "../Components/user-management/UserListItem";
 import { BarLoader } from "react-spinners";
 import { useTranslation } from "react-i18next";
@@ -11,7 +11,7 @@ const defaultPagination = {
   limit: 10,
   sortBy: "surname",
   sortDesc: "false",
-  surnameContains: null,
+  searchedField: null,
 };
 
 export default function UserManagementPage() {
@@ -24,7 +24,7 @@ export default function UserManagementPage() {
       paginationSettings.limit,
       paginationSettings.sortBy,
       paginationSettings.sortDesc,
-      paginationSettings.surnameContains,
+      paginationSettings.searchedField,
     ],
     queryFn: () =>
       getUsers(
@@ -32,7 +32,7 @@ export default function UserManagementPage() {
         paginationSettings.limit,
         paginationSettings.sortBy,
         paginationSettings.sortDesc,
-        paginationSettings.surnameContains
+        paginationSettings.searchedField
       ),
   });
 
@@ -45,6 +45,37 @@ export default function UserManagementPage() {
           pagination={paginationSettings}
           setPagination={setPaginationSettings}
           availablePageNumber={data?.totalPages}
+          limitObjectName="users"
+          sortFieldOptions={
+            <>
+              <option value="name">{t("PaginationSettings.fieldName")}</option>
+              <option value="surname">
+                {t("PaginationSettings.fieldSurname")}
+              </option>
+              <option value="email">
+                {t("PaginationSettings.fieldEmail")}
+              </option>
+              <option value="role">{t("PaginationSettings.fieldRole")}</option>
+              <option value="maxTotal">
+                {t("PaginationSettings.fieldMaxTotalEntries")}
+              </option>
+              <option value="maxSinglePhotos">
+                {t("PaginationSettings.fieldMaxSingles")}
+              </option>
+              <option value="maxCollections">
+                {t("PaginationSettings.fieldMaxCollections")}
+              </option>
+              <option value="isEnabled">
+                {t("PaginationSettings.fieldIsEnabled")}
+              </option>
+              <option value="isNonLocked">
+                {t("PaginationSettings.fieldIsNonLocked")}
+              </option>
+            </>
+          }
+          searchByFieldName="surname"
+          firstPage={data?.firs}
+          lastPage={data?.last}
         />
         <div className="hidden xl:grid xl:grid-cols-9 px-3 py-5 font-bold text-lg text-slate-700 bg-white mt-2 rounded-md shadow">
           <p>{t("PaginationSettings.fieldName")}</p>
