+++ javinukai-front/src/services/records/getCompetitionRecord.js
