+++ javinukai-front/src/services/records/getCompetitionRecords.js
@@ -1,0 +1,21 @@
+export default async function (contestId, categoryId) {
+  const res = await fetch(
+    `${
+      import.meta.env.VITE_BACKEND
+    }/api/v1/records/contest/${contestId}/category/${categoryId}`,
+    {
+      method: "GET",
+      mode: "cors",
+      cache: "no-cache",
+      credentials: "include",
+      headers: {
+        "User-Agent": "react-front",
+      },
+    }
+  );
+  if (!res.ok) {
+    const err = await res.json();
+    throw new Error(err.title);
+  }
+  return await res.json();
+}
